--- src/app/(after-login)/dashboard/default.tsx
@@ -1,3 +1,0 @@
-export default function Default() {
-  return '엄준식입니다';
-}
